refactor: move try_only_named.. to support
